Get it working
- Fix ArticleHandler. This was resetting articles too much on launch
- Fix pagerState. This needs to wait for async initialization for when the
  findIndex is called on the current article
- Remove extra IndexedArticle handling. This caused articles to be
  out-of-date
- Remove up-down pagination. This is additional complexity that will
  need to be stripped out entirely

+++ app/src/main/java/com/capyreader/app/ui/articles/ArticleHandler.kt
@@ -3,15 +3,24 @@ package com.capyreader.app.ui.articles
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import com.capyreader.app.common.AppPreferences
+import com.jocmp.capy.Article
+import com.jocmp.capy.logging.CapyLog
 import org.koin.compose.koinInject
 
 @Composable
 fun ArticleHandler(
+    article: Article?,
     appPreferences: AppPreferences = koinInject(),
     onRequestArticle: (articleID: String) -> Unit,
 ) {
-    LaunchedEffect(Unit) {
+    LaunchedEffect(article?.id) {
+        if (article != null) {
+            return@LaunchedEffect
+        }
+
         val articleID = appPreferences.articleID.get()
+
+        CapyLog.info("handler", mapOf("article_id" to articleID))
 
         if (articleID.isNotBlank()) {
             appPreferences.articleID.delete()
